python/requirements.txt
pyqt6
PyQt6-WebEngine
zmq
colorlog
numpy
scipu
pyvistaqt
ouster-sdk
pypylon
scipy
pyserial
PyQtWebEnginewebsockets
websockets
vedo